demo/requirements.txt
python-dotenv==1.0.1
-i https://pypi.tuna.tsinghua.edu.cn/simple
